+++ PriceAggregator.Core/IRepository/IFeaturedItemRepository.cs
@@ -6,7 +6,7 @@ namespace PriceAggregator.Core.IRepository;
 public interface IFeaturedItemRepository : IRepository<FeaturedItem>
 {
     Task<IEnumerable<FeaturedItem>> GetFeaturedItems(Expression<Func<FeaturedItem, bool>> filter = null, bool isTracked = true);
-    Task<FeaturedItem> AddFeaturedProducts(int[] productIds, string name);
-    Task<bool> DeleteFeaturedItem(int id);
-    Task<FeaturedItem> UpdateFeaturedProduct(int id, string name, int[] productIds);
+    Task<FeaturedItem> AddFeaturedProducts(long[] productIds, string name);
+    Task<bool> DeleteFeaturedItem(long id);
+    Task<FeaturedItem> UpdateFeaturedProduct(long id, string name, long[] productIds);
 }
